block修改后其余需要修改的地方，以及补充了nodestate数据库方面的使用，客户端绑定了本地，考虑进行网络的测试

+++ blockchain/src/main/java/edu/ictt/blockchain/socket/pbft/queue/AbstractVoteMsgQueue.java
@@ -105,7 +105,7 @@ public abstract class AbstractVoteMsgQueue extends BaseMsgQueue {
 	/**
      * 清理旧的block的hash
      */
-	protected void clearOldBlockHash(int number){
+	protected void clearOldBlockHash(long number){
 		TimerManager.schedule(()->{
 			for(String key : voteMsgConcurrentHashMap.keySet()){
 				if(voteMsgConcurrentHashMap.get(key).get(0).getNumber()<=number){
